Move all base view model classes to MVVM root.
