+++ src/StudentToolkit/MVVM/BaseViewModels/DialogViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace StudentToolkit.MVVM.ViewModels.Base;
+﻿namespace StudentToolkit.MVVM.BaseViewModels;
 
 public abstract class DialogViewModel : ValidatableViewModel
 {
@@ -7,7 +7,7 @@ public abstract class DialogViewModel : ValidatableViewModel
         CloseDialogCommand = new DelegateCommand(() => Close?.Invoke());
     }
 
-    public Action? Close{ get; set; }
+    public Action? Close { get; set; }
 
     public virtual ICommand CloseDialogCommand { get; }
 }
