+++ src/StudentToolkit/MVVM/Main/ViewModels/MainViewModel.cs
@@ -1,6 +1,4 @@
-﻿using StudentToolkit.MVVM.About.ViewModels;
-
-namespace StudentToolkit.MVVM.Main.ViewModels;
+﻿namespace StudentToolkit.MVVM.Main.ViewModels;
 
 public class MainViewModel : ViewModel
 {
