+++ tests/StudentToolkit.WPF.UnitTests/GlobalUsings.cs
@@ -1,7 +1,7 @@
 global using SimpleInjector;
 
-global using StudentToolkit.MVVM.Models.Navigation.Messages;
-global using StudentToolkit.MVVM.ViewModels.Base;
+global using StudentToolkit.Navigation.Messages;
+global using StudentToolkit.MVVM.BaseViewModels;
 global using StudentToolkit.WPF.UnitTests.TestDoubles.Dummy.ViewModels;
 global using StudentToolkit.WPF.UnitTests.TestDoubles.Stubs.ViewModels;
 global using StudentToolkit.WpfCore.Services;
